compiler, ide: 64 bit fixes to get the IDE working - Recomputing Class structSize for 'class_fixed' classes in the compiler

+++ compiler/bootstrap/include/stdlib.h
@@ -1,7 +1,7 @@
 #ifndef BSL_STDLIB_H
 #define BSL_STDLIB_H
 
-typedef unsigned int size_t;
+typedef uintptr size_t; // For now, until we add uintsize
 
 /*
 #include <stddef.h>
